(BSR) fix(AuthContext): require cycle
